[ALPS05038609] v4l2: enable IOMMU
Early porting test for removing ION fd path
VP/VR/SVP test pass

MTK-Commit-Id: df2b509fc563b71d736a4bc74ab03895c7d9ee05

CR-Id: ALPS05038609
Feature: [Module]Video Driver

+++ include/uapi/linux/mtk_vcu_controls.h
@@ -30,7 +30,7 @@
  * @va: kernel virtual address
  */
 struct mem_obj {
-	u32 iova;
+	u64 iova;
 	u32 len;
 	u64 pa;
 	u64 va;
@@ -186,7 +186,7 @@ enum gce_event_id {
 
 #if IS_ENABLED(CONFIG_COMPAT)
 struct compat_mem_obj {
-	u32 iova;
+	u64 iova;
 	u32 len;
 	compat_u64 pa;
 	compat_u64 va;
